refactor: migrate JSDoc and AST parsing logic into modular library utilities for documentation validation
